requirements.txt
langchain 
langgraph
langchain-community
langchain-text-splitters
python-dotenv
PyPDF2
beautifulsoup4
html2text
pandas
chromadb
streamlit